+++ cocos2d/CCScheduler.h
@@ -148,7 +148,7 @@ struct _hashUpdateEntry;
  The lower the priority, the earlier it is called.
  @since v0.99.3
  */
--(void) scheduleUpdateForTarget:(id)target priority:(int)priority paused:(BOOL)paused;
+-(void) scheduleUpdateForTarget:(id)target priority:(NSInteger)priority paused:(BOOL)paused;
 
 /** Unshedules a selector for a given target.
  If you want to unschedule the "update", use unscheudleUpdateForTarget.
