+++ cocos2d/CCSpriteBatchNode.m
@@ -205,7 +205,7 @@ static SEL selSortMethod =NULL;
 }
 
 // override addChild:
--(void) addChild:(CCSprite*)child z:(int)z tag:(int) aTag
+-(void) addChild:(CCSprite*)child z:(NSInteger)z tag:(NSInteger) aTag
 {
 	NSAssert( child != nil, @"Argument must be non-nil");
 	NSAssert( [child isKindOfClass:[CCSprite class]], @"CCSpriteBatchNode only supports CCSprites as children");
@@ -217,7 +217,7 @@ static SEL selSortMethod =NULL;
 }
 
 // override reorderChild
--(void) reorderChild:(CCSprite*)child z:(int)z
+-(void) reorderChild:(CCSprite*)child z:(NSInteger)z
 {
 	NSAssert( child != nil, @"Child must be non-nil");
 	NSAssert( [children_ containsObject:child], @"Child doesn't belong to Sprite" );
@@ -265,7 +265,7 @@ static SEL selSortMethod =NULL;
 {
 	if (isReorderChildDirty_) 
 	{	
-		int i,j,length=children_->data->num;
+		NSInteger i,j,length=children_->data->num;
 		id* x=children_->data->arr;		
 		id tempItem;
 		CCSprite *child;
@@ -307,7 +307,7 @@ static SEL selSortMethod =NULL;
 {
 	CCArray *array = [sprite children];
 	NSUInteger count = [array count];
-	int oldIndex;
+	NSInteger oldIndex;
 	
 	if( count == 0 )
 	{	
@@ -496,7 +496,7 @@ static SEL selSortMethod =NULL;
 }
 
 
--(NSUInteger)atlasIndexForChild:(CCSprite*)sprite atZ:(int)z
+-(NSUInteger)atlasIndexForChild:(CCSprite*)sprite atZ:(NSInteger)z
 {
 	CCArray *brothers = [[sprite parent] children];
 	NSUInteger childIndex = [brothers indexOfObject:sprite];
